app/requirements.txt
# app/requirements.txt

Flask==2.2.2
Werkzeug<3.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0